requirements.txt
gooddata-sdk~=1.1
gooddata-pandas~=1.1
python-dotenv~=0.21.0
tqdm~=4.64.1
statsmodels~=0.13.5
seaborn~=0.12.1